:truck: Move files to related folder, redis
